Delete a line that does nothing
`new_dict = {k:v for k,v in another_dict.iteritems()}` makes a copy of `another_dict`,
but nothing more. This is a dictionary comprehension. Just like a list comprehension
results in a list, a dictionary comprehnsion results in a dictionary. But this
comprehension says "make a new dictionary with every key and value from another
dictionary", which simply makes a copy. That is not needed here.